Refactor: Split PlatformAnalyticsPage into components
Split PlatformAnalyticsPage.tsx into smaller, reusable components to improve maintainability and readability.
