+++ src/components/admin/analytics/UserRoleDistributionChart.tsx
@@ -1,0 +1,56 @@
+
+import React from 'react';
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { ChartContainer, ChartTooltip, ChartTooltipContent } from "@/components/ui/chart";
+import { PieChart, Pie, Cell } from "recharts";
+
+interface UserRoleData {
+  role: string;
+  count: number;
+  percentage: number;
+}
+
+interface UserRoleDistributionChartProps {
+  data: UserRoleData[];
+}
+
+const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042', '#8884D8'];
+
+export const UserRoleDistributionChart: React.FC<UserRoleDistributionChartProps> = ({ data }) => {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>User Role Distribution</CardTitle>
+      </CardHeader>
+      <CardContent>
+        <ChartContainer
+          config={{
+            count: {
+              label: "Users",
+              color: "#8884d8",
+            },
+          }}
+          className="h-[300px]"
+        >
+          <PieChart>
+            <Pie
+              data={data}
+              cx="50%"
+              cy="50%"
+              labelLine={false}
+              label={({ role, percentage }) => `${role}: ${percentage}%`}
+              outerRadius={80}
+              fill="#8884d8"
+              dataKey="count"
+            >
+              {data.map((entry, index) => (
+                <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
+              ))}
+            </Pie>
+            <ChartTooltip content={<ChartTooltipContent />} />
+          </PieChart>
+        </ChartContainer>
+      </CardContent>
+    </Card>
+  );
+};
